Commited a working project
The project is working as it is, when user writes a code on the text input
and presses the button it shows the QR on the next page. Now i added
a html file to display the QR (pass the QR image as a parameter) and
add a back button to go back to the initial form.